Don't use AccountAccess directly
Accounts are migrated to NoteDb and then ReviewDb#accounts() will be
removed. The new way to read accounts is by using the Accounts class
which makes it transparent for the caller whether accounts are read from
ReviewDb or NoteDb.

FindOwnerIT is adapted to use Accounts to load accounts for
verification.

Since change I363a8ee040 in Gerrit core the account cache is available
to Prolog predicates. Checker is now using the account cache to lookup
accounts. This should also result in better performance than loading the
accounts from the database each time Prolog rules are evaluated.

Change-Id: I92751e401ef3fcb768e315dad4752f35c6a62919
Signed-off-by: Edwin Kempin <ekempin@google.com>