FindOwnersIT: Use AccountCache#maybeGet instead of AccountCache#get
If an account is missing AccountCache#get returns an empty AccountState
to represent the missing account. The test only retrieves an account
from the cache that it assumes exists. If the account is missing it's
better that the test is failing immediately than obfuscating the error
by getting an empty account.

Change-Id: I705dcb2ee00f020f579c87ec2be068c126e551c8
Signed-off-by: Edwin Kempin <ekempin@google.com>