Update patch set 1
Patch Set 1:

> Could you add some comment in code to document the cases when a CL
> does not have email address for the owner and the reviewers?
> I am not sure how those cases should affect owner-approval checks.

In Gerrit you can have accounts without email addresses. These accounts can log in into the WebUI and do reviews. So it's not safe to assume that all reviewers have an email (note that this is different for the Gerrit installations at Google, at Google users must have an email in order to login).

The emails are stored as external IDs. User may choose one of them as preferred email. This is the email they prefer to be contacted through. But they may also choose to not set a preferred email. In this case the preferred email is null and they can't be contacted by email. Users may change or unset their preferred email any time through the REST API.

In order to push changes an email (stored as external ID) is required, but having a preferred email set is not required. This means for change authors the preferred email can be null.

It's also possible that change authors don't have any email at the moment you do the check. They could upload a change and then delete their emails afterwards.

Also when an account is missing the preferred email is null (also see the follow-up change). Normally accounts should never be missing since we don't support deletion of accounts (but it can definitely happen that a user branch gets lost due to manual deletion or bugs).

So since the preferred email is optional and can be unset anytime relying on the preferred email is error-prone.

Although there are plenty of possibilities how the preferred email can be null, you may still be able to use it for this plugin, as probably almost all users (at least at Google) do have a preferred email set and if this plugin misses a user it's also not super-critical.

This change is really about making it obvious in the code that the preferred email can be null. The current code looks like it assumes that the preferred email is never null, which is wrong.

Patch-set: 1
Reviewer: Edwin Kempin <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>