Update patch set 1
Patch Set 1:

> Two other CLs are replacing get with maybeGet,
> but this one is opposite.
> Why are they different?
> Could either get or maybeGet be used in all places?

Sorry about this being confusing.

The account cache has 2 methods to retrieve accounts:

- get: Always returns an AccountState. If the account is missing you get an empty AccountState instance.
- maybeGet: Returns an Optional<AccountState>. If the account is missing you get an empty Optional.

Almost none of the current callers of the get method is aware that an empty AccountState instance is returned if the account is missing. E.g. in this plugin you expect that the account of the returned AccountState has a preferred email, but for the empty AccountState instance the preferred email is null. Most callers really want and should use maybeGet and handle the missing account case explicitly. This is why we migrate them from using get to using maybeGet (for the find-owners plugin this is done in the predecessor changes).

Now we want to prevent that new callers repeat this mistake of using get. This is why we rename the get method to getEvenIfMissing. This name is more obvious and callers can decide if this is really what they want (this rename is done in change [1]).

Now after get was renamed to getEvenIfMissing, we can rename maybeGet to get, which is done in change [2].

For the find-owners plugin this means:
1. migrate all callers of get to maybeGet (predecessor changes)
2. adapt all callers to the rename of maybeGet to get (this change)

So all the callers in the find-owners plugin consistently use the new get method which was previously named maybeGet.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/154472
[2] https://gerrit-review.googlesource.com/c/gerrit/+/154474

Patch-set: 1
Reviewer: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>