Update patch set 1
Patch Set 1:

Could you consolidate all changes to get and maybeGet into one CL?
Or if you need two stage changes, one CL to change get to maybeGet
and another one later from maybeGet to get.

I think the change of 'get' semantic is really dangerous.
If it is not too late, I would rather Gerrit core not to make that kind of change or require plugins to make changes out and back to 'get'.

Patch-set: 1