Update patch set 1
Patch Set 1:

I didn't know that Gerrit allows accounts without email address and still be a reviewer or author of a change. Is there a possibility to disallow that from Gerrit core?

Did you see the find-owners plugin crashed in those cases?
If there was no crash, either the cases did not happen or
it was just lucky that null keys are used in those maps.

Could you just add short comments in the code to mention that Gerrit allows accounts without email addresses? For those accounts, their +1 or default +1 won't be counted as owner approval because all OWNERS files have are email addresses.
Thanks.

Patch-set: 1