Update patch set 2
Patch Set 2:

> Could you consolidate all changes to get and maybeGet into one CL?

Sure, no problem. Done.

> Or if you need two stage changes, one CL to change get to maybeGet
> and another one later from maybeGet to get.

Two stages are not needed.

> I think the change of 'get' semantic is really dangerous.

The old 'get' method is still available as 'getEvenIfMissing'. If you prefer you can stick with using this method. 
But as explained in the commit message using the new 'get' method (which was named 'maybeGet' before) doesn't change much of the behaviour.

> If it is not too late, I would rather Gerrit core not to make that kind of change
> or require plugins to make changes out and back to 'get'.

If you think that this is better, alternatively approve this change [1], which only adapts to the rename of 'get' to 'getEvenIfMissing'. The semantics of this method isn't changed, it only gets a new name.

[1] https://gerrit-review.googlesource.com/c/plugins/find-owners/+/156811

Patch-set: 2
Reviewer: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>