Update patch set 3
Patch Set 3:

> The renaming plan really sounds confusing to users.
> If I understand it correctly, all Gerrit core users,
> plugin writers will need to
>   change 'get' to 'getEvenIfMissing'
>   change 'maybeGet' to 'get'
> at once when Gerrit core changes.
> What would break if anything is not changed at the same time?

No. The change that renames this method will be part of the next Gerrit release. For this release there will be several release candidates before the final release. This gives the plugin owners time to test their plugin against the new Gerrit version and do adaptions if needed. Of course plugin owners may regular build their plugins against Gerrit master to notice breakages early. But a breakage at this point doesn't matter since nobody runs the new Gerrit version before a release is done.

It is unique to Google that we consume changes from the master branch almost instantly. This is why I uploaded changes for all plugins that are used at Google that adapt them to the rename. In Gerrit these changes share the same topic with the change in Gerrit core that does the rename. Hence these changes will all be submitted together. I will then take care to import them together and then they are rolled out together.

> So this CL don't rename 'get' but want to adopt its new meaning,
> as old maybeGet, right?

Yes, this is correct.

> I don't think the extra handling of null preferred email actually
> changes find-owner behavior to users. With or without this change,
> accounts without preferred email are not counted into owner approval
> system because null or missing cannot match any existing email address
> in OWNERS files.

Yes, exactly. This is what I tried to explain in the commit message :)

> On the other hand, if Gerrit core allows an account to have email address
> but not 'preferred' email, why don't we change the getPreferredEmail function
> to return the 1st email address? This makes the 1st email address as the
> 'preferred' one.

We cannot easily change the behaviour of the getPreferredEmail method, but you can implement this fallback yourself. The emails are stored in the external IDs. You can get the external IDs by AccountState#getExternalIds() and then you can use ExternalId#getEmails(Collection<ExternalId>) to extract the emails from the external IDs. Implementing this fallback is a good idea and I would recommend to do it.

> Isn't that the current behavior when a new user signs in?

For Google we definitely set the preferred email when creating new accounts. But since this is an open source plugin, it shouldn't rely on Google specifics. For open source Gerrit installations it is not guaranteed that a preferred email is set when a new account is created.

> We need to handle accounts without any 'email' address,
> but I think all those cases should be caught by exceptions.

Yes, I think so too. But in this change I'm only doing the minimal changes to not break the plugin when the rename done.

> The code in OwnersDb.java:155 is to translate an email address found in OWNERS files
> to their 'preferred' form. If preferred form is not found, we should just use
> the original form in 'owner'.

SGTM, but this should be done in follow-up changes.

Patch-set: 3