Create "Find Owners" button after showing a page.
When switching PolyGerrit pages, the "Find Owners" button
could be hidden or lost. Previous change to un-hidden the
button does not work now. So we always create a new button
and remove the old button.

Change the pop up window message to be more generic.
Some Gerrit servers are configured with this plugin
but do not enforce the owner approval rule in pre-submit checks.

Change-Id: I09fc432708df97ce248443586d081e39f278b41f