Update patch set 2
Patch Set 2:

(3 comments)

Nice work, Chris!

I agree to chh@ that the two systems need to be able to co-exist, but I think this is easy to achieve when adding a config option as per my in-line comment.

I don't think you need to implement all the features in a single change, but that this can be iterative, especially if the submit rule is gated behind a config option.

Patch-set: 2
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>