Format all files with google-java-format 1.7
This ensures that files obey the formatting rules in the Google Java
Style Guide[1]. Moreover, it matches the formatting to the Gerrit core
project. This has benefits such as making refactoring across plugins and
core easier, since developers can use the same tooling regardless of
which repo they happen to be looking at at the moment.

[1] https://google.github.io/styleguide/javaguide.html

Change-Id: Ie5a0ca3230a6156f20f370ffac0b0b307eae8f07