Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I don't mind adding more finer log messages.
> If the stats have hit ratio and space info,
> that might be useful to tune up caching parameters depending on Gerrit server.
> 
> The cache is effective and needed for single process Gerrit server,
> which can issue multiple (http) requests of the owners info for one user action.
> It's less effective when multiple requests might hit different servers.

When looking at the cache I was wondering if we can just remove it. 30s object lifetime and the fact that two requests might not hit the same backend on our deployment while the cache is in-memory only hinted in that direction. Metrics should proof if this is correct.

I am wondering though, what is the request pattern here? I see how this is called when evaluating submit rules (=once when we show the change page). Deciding if the action should be added to the action menu does not need owners evaluation, so no cache call. The only call I see is to the find-owners REST endpoint which the user might not use at all (gut-feeling says 90% of change screen views never call the action as a follow-up).

Patch-set: 3