Update patch set 3
Patch Set 3:

It will be easier to see the effect with a local dev/test Gerrit server,
and measure/dump every call to get OwnersDb. If I remember it correctly, a click
on [FIND OWNERS] button, or change of a few reviewers, or vote on the label,
can trigger multiple "submittability" checks and thus asking for OwnersDb.
Sometimes it's multiple http requests, sometimes it's one.

Do you know the overhead of caching interface,
when the caching duration is default 0?
The OWNERS file is getting more complicated and creating OwnersDb more expensive.
Having a caching mechanism option should be better.

Patch-set: 3