Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> It will be easier to see the effect with a local dev/test Gerrit server,
> and measure/dump every call to get OwnersDb. If I remember it correctly, a click
> on [FIND OWNERS] button, or change of a few reviewers, or vote on the label,
> can trigger multiple "submittability" checks and thus asking for OwnersDb.
> Sometimes it's multiple http requests, sometimes it's one.
> 
> Do you know the overhead of caching interface,
> when the caching duration is default 0?
> The OWNERS file is getting more complicated and creating OwnersDb more expensive.
> Having a caching mechanism option should be better.

There is one place where Gerrit core would evaluate submit rules twice while once is enough but I2d77cf93b5 fixed that. Apart from that, each request should trigger one rule execution at maximum. The same goes for a request to the Action.

> Do you know the overhead of caching interface,
> when the caching duration is default 0?

The overhead is near-zero performance-wise, but when looking at this I wasn't concerned about the overhead of caching performance-wise, rather the fact that we cache mutable state and pointers to service objects through OwnersDb and key that by another service object statically per-JVM.

For maintainability purposes we try to cache only immutable data objects in Gerrit (for example to avoid races) and hook up caches centrally for monitoring and serialization. So if the cache is not really needed for performance reasons, I think it would be a good idea to remove it. Then we wouldn't have to worry about it's structure.

The alternative is to hook it up to Gerrit's cache infrastructure and make the key/values immutable. I have a draft change for this, but it touches a lot of the plugin's code so I am worried of accidentally breaking something and would go down that route only if we really have to.

Patch-set: 3