Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (8 comments)
> 
> Nice, thanks for doing this.
> 
> My biggest concern is how we are going to test this, given that this is part of the backend rollout. Do you have a plan?

Ya thats my concern as well as we don't have any tests for this plugin as well. For testing:
0. Add some test cases
1. I added chh@ as one of the reviewer to make sure the code logic is still fine and covers what we had
2. I am thinking of having it released to canary and test manually for polymer 1 (as no polymer 2 in canary that has find-owners plugin enabled), and all hosts with find-owners in prod are still in polymer 1
3. Test on polymer 2 on prod with `polymer2-` mirrors before enabling p2 for them

Another alternative that will be safer but much more work:

we have the new code as a new plugin, and then replace it on one host, and then same steps as above, and we slowly replace it on all hosts or just merge it back to find-owners.

Patch-set: 1