Add configuration option to set behavoir of Author commits.
We use the find-owners plugin to enforce Owner +2 votes.
Currently this fails when a change Author is an owner
since the code assumes the author approves the change and
then any +2 vote would allow the code to be submitted.
This change adds a configuration option to set the desired
behavior. The default value keeps the current behavior.
Setting the option to false makes no assumptions about an
Author vote and changes will not be able to be submitted
unless an owner votes on the change.

Change-Id: I6d8230c4f12fe93cfb76bbcee5f74c6940181654