findowners: Fix concurrency issue in Config class
The Config class seems to be used in multi-threaded contexts (e.g. via
Cache -> OwnersDb) but is not hardened for multi-threading. We noticed
errors in our logs complaining about concurrent modifications of
projectConfigMap when computeIfAbsent() is called on it. Fix that
immediate issue by switching to a concurrent type for the map.

Change-Id: I1b3f2ae7211787974440444c12c2c703e9d39d9b