Adapt tests to compact JSON
Since change I2f1cb91e9 Gerrit now always returns compact JSON. Adapt
the test to this.

Instead of doing string assertions on the JSON the find-owners tests
should rather parse the JSON with GSON and then do assertions on the
parsed objects so that the tests would not be affected by such a change.
However since the find-owners plugin is deprecated (soon to be replace
by the code-owners plugin) we don't want to invest time in rewriting
these tests.

ApiIT.getOwnersTest was already failing before, as the owner_revision
field was not set in the expected response. While we are here, fix this
too.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I616b3fe74ac99fafdd986442751eb74c69c0ac5c
(cherry picked from commit fabbbf05f30fb1bfd5c43f6065db5c6b6ffe7b31)