Update link to the code-owners plugin
So far the link was pointing to the code-owners plugin documentation in
Gitiles. At this place the source code of the documentation can be seen,
but links in it do not work there.

Instead point to the hosted plugin documentation of the android host
which has this plugin installed. There we have the rendered
documentation and links work here.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ic7620a01dcf40fbd15543f6d9f19952d512aa83d