Adapt to changes in gerrit core
Change I78c15e01 migrated PrologEnvironment off of PatchListCache.
find-owners used to reference this field which is no longer valid, hence
fixing with this change. This field was redundant in find-owners anyway,
so we are just removing it.

While we are at it, cleanup some unused code in find-owners as well.

Change-Id: I31d938875d857775263dc17e8e1f8b5b47f0f41d